begun spliting reducers
